FastAPI_requirements.txt
aiosqlite==0.19.0
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==3.2.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.2
databases==0.9.0
defusedxml==0.7.1
dnspython==2.4.2
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.0
filelock==3.13.1
fsspec==2023.12.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httpie==3.2.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
jose==1.0.0
Mako==1.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
networkx==3.2.1
orjson==3.9.10
passlib==1.7.4
pip==24.0  # Please note that the version of pip is not typically included in requirements.txt
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
pydantic==2.5.3
pydantic-core==2.14.6
pydantic-extra-types==2.4.0
pydantic-settings==2.1.0
Pygments==2.17.2
PyJWT==2.8.0
pyngrok==7.1.2
PySocks==1.7.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
requests-toolbelt==1.0.0
rich==13.7.0
rsa==4.9
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.27
starlette==0.35.1
sympy==1.12
torch==2.1.2
typing-extensions==4.9.0
ujson==5.9.0
urllib3==2.1.0
uvicorn==0.26.0
watchfiles==0.21.0
websockets==12.0